backend/requirements.txt
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
feedparser==6.0.2
Flask==1.1.2
fuzzysearch==0.7.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pdfminer==20191125
pdftotext==2.1.5
psycopg2==2.8.6
pycodestyle==2.6.0
pycryptodome==3.9.9
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
schedule==1.0.0
sgmllib3k==1.0.0
SQLAlchemy==1.3.22
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
sklearn==0.0
